hugo_migration/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Markdown==3.5.1
Pygments==2.17.2
python-dotenv==1.0.0
PyYAML==6.0.1
GitPython==3.1.40
python-frontmatter==1.0.8